retrieve view results in batches of 1k, add count function
Change implementations of `apply_view` and `apply_temp_view` in a way that
they only retrieve a limited number of view results at a time. They still go through
the entire view result row list, but they won't download a huge bunch of data
at once.

The new function `view_result_count` queries a collection with a given view, but
without downloading any results. Instead, it just returns the number of results to
be expected if the view would be executed as usual.